Relicense to Artistic 2.0 (same as Parrot)
Previously Blizkost had no license at all, not a nice condition for an
open-source project to be in.  I have e-mails and IRC approval from all
committers in 'CREDITS' and 'git log'.  The Parrot license was chosen
because Blizkost's integration with Parrot is far deeper than its
integration with Perl 5.